Ok now for real, fix orders (#14)
* Scale issues fixed

Only one last tiny bug with orders that need to be handled (shouldn't be able to add more ingredients to complete orders).

* Created a bun prefab

The bun now spawns in the pantry and a burger is now made of bread + patty + cheese

* Orders 100% fixed

Player can now combine orders with other orders, can't add ingredients to complete orders anymore, and Ingredient and Order classes are more organized (saying the Order class is organized is a bit of a stretch, but it's MORE organized than before, which I think is a victory).

---------

Co-authored-by: tiffne <105519621+tiffne@users.noreply.github.com>